requirements.txt
Django>=2.0,<2.1
pytz==2016.7
psycopg2==2.7.3.2
djangorestframework==3.7.7
